Update and rename unix_and_perl_project_0.pl to unix_and_perl_project_5.pl
This file will return the log odds of every k-mer for the intronic sequences in the IME_intron sequence file. Execute this file in the terminal. It will ask for the value of k. After that it will return the list of k-mers, with their frequencies, odds ratio and log-odds ratio.